star64: include only necessary patches for the kernel
When looking at the patches in the default SRC_URI, 2 patches
are already included in the star64 kernel, and 1 patch is not
applicable (as the problem that is fixed is not present in the
fork). Due to this the build failed during do_patch task.

This change removes the unneeded patches from star64 board,
and also updates the kernel revision to the latest available one.

Also, a typo is fixed in the kas config file name.